Ignore user-created project files.
This is sortof a lazy way to let people experiment without messing up
the repo history... it just removes their ability to use version control
for their own work. I think this is fine, because the goal here is to do
the exercises on your own, and the repo is just tooling to help with
that.